Moved componentloader to general ui package
