+++ Code/src/main/java/nl/utwente/viskell/ui/ComponentLoader.java
@@ -1,4 +1,4 @@
-package nl.utwente.viskell.ui.components;
+package nl.utwente.viskell.ui;
 
 import javafx.fxml.FXMLLoader;
