+++ Code/src/main/java/nl/utwente/viskell/ui/components/ArgumentSpace.java
@@ -12,6 +12,7 @@ import javafx.scene.input.TouchEvent;
 import javafx.scene.layout.Pane;
 import javafx.scene.layout.Region;
 import javafx.scene.shape.Circle;
+import nl.utwente.viskell.ui.ComponentLoader;
 import nl.utwente.viskell.ui.ConnectionCreationManager;
 
 import java.util.ArrayList;
