+++ Code/src/main/java/nl/utwente/viskell/ui/components/DefinitionBlock.java
@@ -11,6 +11,7 @@ import nl.utwente.viskell.haskell.expr.Lambda;
 import nl.utwente.viskell.haskell.expr.LocalVar;
 import nl.utwente.viskell.haskell.type.FunType;
 import nl.utwente.viskell.haskell.type.Type;
+import nl.utwente.viskell.ui.ComponentLoader;
 import nl.utwente.viskell.ui.CustomUIPane;
 
 import java.util.ArrayList;
