+++ Code/src/main/java/nl/utwente/viskell/ui/components/InputArgument.java
@@ -9,6 +9,7 @@ import javafx.scene.control.Label;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.Pane;
+import nl.utwente.viskell.ui.ComponentLoader;
 
 /**
  * A class that represents an input field inside a FunctionBlock.
